requirements.txt
streamlit>=1.25
tensorflow>=2.10
numpy
Pillow
